Restructured code. Non-essential functionality for creating scan sequences are now in BCSimConvenience. Updated client code
