+++ src/ScanSequence.cpp
@@ -32,7 +32,6 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "bcsim_defines.h"
 #include "ScanSequence.hpp"
 #include "vector3.hpp"
-#include "rotation3d.hpp"
 
 namespace bcsim {
 
@@ -100,111 +99,4 @@ bool ScanSequence::is_valid() const {
     return res;
 }
 
-ScanSequence CreateScanSequence(std::shared_ptr<SectorScanGeometry> geometry, size_t num_lines, float timestamp) {
-    const auto line_length = geometry->depth;
-    ScanSequence res(line_length);
-    
-    // Will be transformed by a rotation into lateral and radial unit vectors.
-    // NOTE: Need to use double to avoid "not orthonormal" error (maybe the test is to strict for floats?)
-    const auto unit_vector_x = unit_x<double>();
-    const auto unit_vector_z = unit_z<double>();
-    const vector3 origin(0.0f, 0.0f, 0.0f);
-
-    for (int line_no = 0; line_no < static_cast<int>(num_lines); line_no++) {
-        const float angle = -0.5f*geometry->width + geometry->tilt + line_no*geometry->width/(num_lines-1);
-
-        const auto ROT_MATRIX = rotation_matrix_y<double>(angle);
-        const auto temp_radial_direction  = boost::numeric::ublas::prod(ROT_MATRIX, unit_vector_z);
-        const auto temp_lateral_direction = boost::numeric::ublas::prod(ROT_MATRIX, unit_vector_x); 
-	
-        // Copy to vector3 vectors. TODO: deduplicate
-        const vector3 direction  ((float)temp_radial_direction(0),  (float)temp_radial_direction(1),  (float)temp_radial_direction(2));
-        const vector3 lateral_dir((float)temp_lateral_direction(0), (float)temp_lateral_direction(1), (float)temp_lateral_direction(2));
-
-        try {
-            auto sl = Scanline(origin, direction, lateral_dir, timestamp);
-            res.add_scanline(sl);
-        } catch (std::runtime_error& e) {
-            throw std::runtime_error(std::string("failed creating scan line: ") + e.what());
-        }
-        
-    }
-    return res;
-}
-
-ScanSequence CreateScanSequence(std::shared_ptr<LinearScanGeometry> geometry, size_t num_lines, float timestamp) {
-    const auto line_length = geometry->range_max;
-    ScanSequence res(line_length);
-
-    const auto unit_vector_x = unit_x<double>();
-    const auto unit_vector_z = unit_z<double>();
-
-    // Copy to vector3 vectors. TODO: deduplicate
-    const vector3 direction  (unit_vector_z(0), unit_vector_z(1), unit_vector_z(2));
-    const vector3 lateral_dir(unit_vector_x(0), unit_vector_x(1), unit_vector_x(2));
-
-
-    for (int line_no = 0; line_no < static_cast<int>(num_lines); line_no++) {
-        try {
-            const vector3 scanline_origin(-0.5f*geometry->width + line_no*geometry->width/(static_cast<int>(num_lines)-1), 0.0f, 0.0f);
-            auto sl = Scanline(scanline_origin, direction, lateral_dir, timestamp);
-
-            res.add_scanline(sl);
-        } catch (std::runtime_error& e) {
-            throw std::runtime_error(std::string("failed creating scan line: ") + e.what());
-        }
-    }
-    
-    return res;
-}
-
-// probe_origin is position of probe's origin in world coordinate system.
-ScanSequence CreateScanSequence(ScanGeometry::ptr geometry, size_t num_lines, float timestamp) {
-    auto sector_geo = std::dynamic_pointer_cast<SectorScanGeometry>(geometry); 
-    auto linear_geo = std::dynamic_pointer_cast<LinearScanGeometry>(geometry);
-    if (sector_geo) {
-        return CreateScanSequence(sector_geo, num_lines, timestamp);
-    } else if (linear_geo) {
-        return CreateScanSequence(linear_geo, num_lines, timestamp);
-    } else {
-        throw std::runtime_error("unable to cast scan geometry");
-    }
-}
-
-namespace detail {
-    // Apply a 3x3 rotation matrix to a vector3;
-    template <typename T>
-    vector3 TransformVector(const vector3& v, const boost::numeric::ublas::matrix<T>& matrix33) {
-        boost::numeric::ublas::vector<T> temp_v(3);
-        temp_v(0) = static_cast<T>(v.x); temp_v(1) = static_cast<T>(v.y); temp_v(2) = static_cast<T>(v.z);
-        const auto transformed = boost::numeric::ublas::prod(matrix33, temp_v);
-        return vector3(static_cast<bc_float>(transformed(0)),
-                       static_cast<bc_float>(transformed(1)),
-                       static_cast<bc_float>(transformed(2)));
-    }
-}
-
-ScanSequence::s_ptr OrientScanSequence(const ScanSequence& scan_seq, const vector3& rot_angles, const vector3& probe_origin) {
-    const auto rot_matrix = rotation_matrix_xyz(rot_angles.x, rot_angles.y, rot_angles.z);
-
-    const auto line_length = scan_seq.line_length;
-
-    auto num_lines = scan_seq.get_num_lines();
-    auto res = new ScanSequence(line_length);
-    for (int i = 0; i < num_lines; i++) {
-        const auto old_line = scan_seq.get_scanline(i);
-        const auto rotated_origin      = detail::TransformVector(old_line.get_origin(),      rot_matrix);
-        const auto rotated_direction   = detail::TransformVector(old_line.get_direction(),   rot_matrix);
-        const auto rotated_lateral_dir = detail::TransformVector(old_line.get_lateral_dir(), rot_matrix);
-        
-        res->add_scanline(Scanline(rotated_origin + probe_origin,
-                                  rotated_direction,
-                                  rotated_lateral_dir,
-                                  old_line.get_timestamp()));
-    }
-    return ScanSequence::s_ptr(res);
-}
-
-
 }   // namespace
-
